code refactoring

+++ src/main/java/ru/daniil/telegrambot/repository/MessageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.daniil.telegrambot.models.Messages;
 
 @Repository
-public interface MessageRepository extends CrudRepository<Messages, Long> {
+public interface MessageRepository extends CrudRepository<Messages, Integer> {
 }
